third_party/blink/web_tests/platform/generic/http/tests/devtools/indexeddb/transaction-promise-console-expected.txt

Ensure transactions created within Promise callbacks are not deactivated due to console activity

PASS: Transaction is still active
